+++ Examen Node.js, MongoDB/src/controllers/peopleController.js
@@ -43,7 +43,7 @@ const peopleController = (People) => {
                 _id: params.id
             })
 
-            if (checkData === null) {
+            if (checkData.lenght === 0) {
                 return res.status(httpStatus.FORBIDDEN).send('No data found with the provided ID.')
             }
 
@@ -87,6 +87,14 @@ const peopleController = (People) => {
     const deletePeopleById = async (req, res, next) => {
         try {
             const { params } = req
+
+            const checkData = await People.find({
+                _id: params.id
+            })
+
+            if (checkData.lenght === 0) {
+                return res.status(httpStatus.FORBIDDEN).send('No data found with the provided ID.')
+            }
 
             await People.findByIdAndDelete(params.id)
